Users#destroy実行制限付与

+++ app/controllers/users_controller.rb
@@ -1,5 +1,7 @@
 class UsersController < Users::ApplicationController
-  before_action :set_user, only: %i(show destroy)
+  include UsersHelper
+
+  before_action :set_user, only: %i(show)
   before_action :set_members, only: %i(index)
   before_action :current_user_admin?, only: %i(new create destroy)
 
@@ -27,9 +29,9 @@ class UsersController < Users::ApplicationController
   def destroy
     $DELETE_COMMAND = "Delete".freeze
     @user = User.find(params[:command])
-    if $DELETE_COMMAND == params[:input_delete]
-      # 現在はユーザーの持ってる案件の完了日が空文字である場合がない場合に分岐させてます
-      if @user.leads.find_by(completed_date: "").empty?
+    if $DELETE_COMMAND == params[:input_delete] && same_company_id_judge(@user)
+      # 現在はユーザーの持ってる案件の完了日が空文字がない場合に分岐させてます
+      if @user.leads.find_by(completed_date: "").blank?
         flash[:success] = "成功しました" if @user.destroy
       else
         flash[:danger] = "未完了の案件を担当しています。別の担当者に変えてください"
